Implement totalhash lookup.
This service can now, optionally, look up a hash on totalhash.com. It
requires an API key to do this.

If the hash exists it will pull down results and store them locally. I've
tried to store as much as I can out of the reports but it's not clear
exactly what can be in a report (I can't find a schema anywhere). I've
looked over many different reports to get a good sample set and this
seems to be pretty accurate to what totalhash will report that we don't
already store via other services (section data, PDB paths, etc).