Move pehash calculation to peinfo from totalhash.
It makes sense to move this to peinfo, we do all our other hashing of PEs
there. It is better suited in peinfo and leaving totalhash just for API calls.

I assigned copyright to Adam for this as this was just a copy/paste of his
code from pehash (which was then put in totalhash and then copied here).