Finish relationships service API.
Remove all the debug stuff and finish this implementation. The major change
in the handler is that we now expose the object type and if it is considered
'visible' or not in the results. It makes it easier for consumers to parse.

Usage of the API looks like this:

http://localhost:8000/api/v1/relationshipsservice/?ctype=Sample&cid=53cdd9e4d6fa25b059c54ddc&depth=1&types=Sample

The ctype and cid parameters are the CRITs type and ID of the starting object.
The depth and types arguments are optional. If depth is not given it defaults
to 3, and if types is not given it defaults to all top-level object types.

As is true of the graph in the UI, if the type of an object does not match the
list in the types parameter it is still collected and included in the graph,
but is marked as 'nvisible. This is so that the path to other objects related
to it is easy to follow.