Fixing TAXII 1.1 and STIX 1.1.1 support.
We now attempt to poll using TAXII 1.1 and if that fails we will inspect
the response to see if the server is trying to tell us that it is a
TAXII 1.0 server. If we determine it is, we will try to poll using TAXII
1.0.

Messages to the user in the interface are more verbose making it easier
to determine what happened.

We now use the appropriate STIX version constants for STIX 1.1.1.

Closes #27, #28.