Fix passivetotal service.
No need to use headers anymore.

"apiKey" is renamed to "apikey" in the submission, so name variable
accordingly.

Don't need to do a json.dumps() on the params anymore as they now MUST be
put in the URI as parameters, not in the body.

The resulting data has also changed slightly, so adjust processing there.