Make it so pysnugglefish doesn't break CRITs.
Due to the way the service was written and added to CRITs, importing
pysnugglefish like this would cause the UI to fail if you didn't install
pysnugglefish first.

This change makes the import localized to the only place it's used so
not having it installed will still allow you to use CRITs.