Provide a way to regenerate WHOIS data.
This isn't really a migration. The script will kick off the WHOIS service
for any existing domain object which has WHOIS data and then move that data
to unsupported_attrs. There is no good way to preserve the existing WHOIS
data because it is now stored in AnalysisResults in a completely different
format.

When the service is run the normal logic applies. That is, if you have a
pyDat URL provided it will use that. If you have a DomainTools username
and API key provided it will also use that. If you have neither it will
do a live lookup.

Also, fix the forms so that default values are provided when the service is
run via the migrate script.