Fix an off-by-one in entropy calc.
If end is -1 it means operate until the end of the data, instead we were
passing it directly as the end which is not inclusive so we were never
including the last byte. This resulted in entropy values which were ever
so slightly off (assuming a large enough sample). Not horrible for
getting a feel for the entropy but bad when you want to copy that value
into a different tool for comparison.