Partial revert of b67fa4c44bcdfd447cde88bcf532b6910d436536.
Upon closer inspection of the pehash paper[1] and looking at the output of
hashes on totalhash and viper I've come to the conclusion that this needs
to be reverted. With this change we are now generating proper hashes, in
line with totalhash and viper.

I know it looks weird but certain bits are intentionally left out, read
the paper for explanation.