STIX Validator Service.
This service adds a Nav item under "Services". You can paste a block of
XML and it will run it through the stix-validator. Any results are
returned to you.

It currently does not support the "Profile" validation as I think it
requires updating a Profile file to do the validation with.