Address concerns with virustotal_lookup.
As discussed in #80, this commit fixes some issues with the improved VT
lookup service.

Move the options to create new PCAP and Domain objects to run-time.

Update any hashes that are on the object with the ones from the report
if they do not match. Given that VT has to have the sample to scan it,
it is safe to assume their hashes are more correct than the ones we
might have.

A lot of whitespace cleanup and general consistent changes. Mostly
making sure that we don't put unnecessary underscores in places and
don't add blank sections in the report.