Fix ANB on events.
A long time ago, in a galaxy far away, collect_objects() was refactored
to do recursion and generally behave much better. This was never brought
up to speed with these changes. :(

This change fixes it so that the behavior is restored while still
relying upon collect_objects() to do the hard work. Now that recursion
is not required here it simplifies the logic in this handler quite a
bit.