Fix a minor issue in whois service.
When the database is being updated the "latest" endpoint will return
nothing until the entry for the new version of the domain exists.

For example, the database has foo.com with version 1, 2, and 3. The
database is then being updated with version 4. While this is going on
the "latest" version is 4 but foo.com may not have been imported yet for
that version. This results in the service telling you which versions
exist in the log but being unable to fetch the most recent one that is
available.

To fix this just walk the results and grab the highest version, which is
the latest _IN_ the database, even while an import is happening.

This also eliminates the need to make multiple requests, which was just
silly anyways.