Campaigns that are related through other TLO's do not render correctly
if multiple campaigns are related through other TLO’s the graph won’t
display correctly.  This is because the first campaign to be added to
the nodes array uses the tlo_styles_dict[‘Campaign’] directly instead
of a copy.  Fix is to copy to a new dictionary based on the
lo_styles_dict.