Update parsers.py
Should fix issues where `header.package_intents` is empty and we end up with "list index out of range". If this throws an exception, we'll now fall back to our default event_type.

Should fix issue with parsing indicator observables and not being able to set `item` because `observable.object_.properties` is not set on the observable. In this case we will now append a failed object for display in the UI similar to how we parse other observables.