Add several little features.
This adds Releasability tracking for exported Indicators. Also converts
the "Type" and "Threat Type" fields on export to be selects. If the
Indicator Type or Threat Type doesn't match one that ThreatExchange will
accept, it will provide a warning letting you know a valid one needs to
be selected and the "Submit" button will be disabled until a valid one
is selected.