Add Related searches to ThreatExchange.
For each result, if there are related searches that can be done, provide
a dropdown to select what kind of related search you want to do. A modal
will pop up with the results. This allows you to do nested related
searches and keep modals up as you go along to see your trail.