Fix Descriptor/MalwareAnalyses imports and Status query field.
Some ThreatExchange Descriptor types aren't in CRITs. Started building a
mapping function to catch these and attempt to find a logical CRITs
Indicator Type for them. If we don't have a mapping we will return an
error that there isn't a supported Type for it. This will help people to
let us know that there's a type missing and what it is.

MalwareAnalyses imports were failing because of a SUBMITTER_COUNT field
not being available on the ThreatExchange side. It's not something we
need so I've added it to the list of fields not to query for.

The Status query field wasn't being used due to a copy/paste id value.