FireEye service to submit samples from CRITs to FireEye MAS appliance via the CMS API.
Currently this service is really basic where it submits a sample and gets the report back. It doesn't parse all of the results back, but it can be scalable to report back more in the future as people see fit. As of today it is returning malware-detected, cnc-services, any os-changes with malicious alert in it (network, files, folders, mutex, process, and regkey).

Couple of things to note, this service was written a while back. I believe with the newer update you can talk to the appliance directly without having to go through the CMS. There is also a v1.1 api version, but my service uses 1.0.

I was hoping to share this service with the community so that it can be improved upon and/or help someone who might find this service userful.