Adding Pretty Things service.
Pretty Things is designed to be a service where people can add graphs,
heatmaps, etc. to visualize data in interesting ways. The initial graph
is a heatmap for Campaign activity over time. It is pretty raw and could
use a bit of touchup, but it functions for my limited dataset.